Create Find Kth Prime
The problem statement is really simple. There are some queries. You are to give the answers.

Input :
An integer stating the number of queries Q(equal to 50000), and Q lines follow, each containing one integer K between 1 and 5000000 inclusive.

Output :
Q lines with the answer of each query: the Kth prime number.